Separate upload from archive record creation
Having upload as a separate function from archive record creation allows
for a wider variety of future ways to upload; for instance an internal
client might upload directly, or in future we might support multi part
upload and want different ways of exposing that flow to clients.

Issue #116
