+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Changed
+- The `createArchiveRecord` function no longer uploads files, clients must use the new `uploadFile` function for this.
+
+## Added
+- The `uploadFile` function now exists.
 
 ## [0.5.4] - 2023-04-18
 ### Fixed
