+++ src/sdk/index.ts
@@ -7,3 +7,4 @@ export * from './getArchiveRecord';
 export * from './getArchives';
 export * from './getAuthenticatedAccount';
 export * from './getFolder';
+export * from './uploadFile';
